gametime is unsigned
svn-id: r5383

+++ gemrb/plugins/GUIScript/GUIScript.cpp
@@ -3883,7 +3883,7 @@ PyDoc_STRVAR( GemRB_GetGameTime__doc,
 
 static PyObject* GemRB_GetGameTime(PyObject * /*self*/, PyObject* /*args*/)
 {
-	int GameTime = core->GetGame()->GameTime;
+	unsigned long GameTime = core->GetGame()->GameTime;
 	return PyInt_FromLong( GameTime );
 }
